poprawa generowania plików DHCP

+++ backend/models/Dhcp.php
@@ -13,13 +13,16 @@ class Dhcp extends Model
 		
 		echo \Yii::getAlias('@console/dhcp');
 		$test = ob_get_contents();
-// var_dump($test); exit();		
-		$sysout = system('rm ' . $test . '/subnets/*');
-		$updateFile = $test . '/subnets/.update_notify';
-// 		var_dump($test); exit();
-		$dhcpSubnets = Subnet::find()->where(['id' => $subnet, 'dhcp' => true])->all();
 		
-		$ipValidator = new IpValidator(['ipv6' => false]);
+		$updateFile = $test . '/subnets/.update_notify';
+// 		$ipValidator = new IpValidator(['ipv6' => false]);
+		
+		if (is_null($subnet)){
+			$dhcpSubnets = Subnet::find()->where(['dhcp' => true])->all();
+			$sysout = system('rm ' . $test . '/subnets/*');
+		} else {
+			$dhcpSubnets = Subnet::find()->where(['id' => $subnet, 'dhcp' => true])->all();
+		}
 		
 		foreach ($dhcpSubnets as $dhcpSubnet){
 			
@@ -28,10 +31,10 @@ class Dhcp extends Model
 			//var_dump($ipValidator->validateAttribute($dhcpSubnet, 'ip')); exit();
 			//if($ipValidator->validateAttribute($dhcpSubnet, 'ip')){
 				
-				$data = "# PODSIEC " . $dhcpSubnet->desc . "\n
-#######################################\n
-#         INTERNET - ADRESACJA\n
-#######################################\n
+				$data = "# PODSIEC " . $dhcpSubnet->desc . "
+#######################################
+#         INTERNET - ADRESACJA
+#######################################
 \n						
 subnet " . $dhcpSubnet->blockIp->getFirstIp() . ' netmask ' . $dhcpSubnet->blockIp->getMask() . " {\n";
 				$options = $dhcpSubnet->generateOptionsDhcp();
@@ -40,26 +43,27 @@ subnet " . $dhcpSubnet->blockIp->getFirstIp() . ' netmask ' . $dhcpSubnet->block
 					$data .= "\t" . $modelDhcpOption->name . " " . $option['value'].";\n";
 				}
 				
-				$data .= "\tdefault-lease-time " . $options[49]['value'] . ";\n
-\tmax-lease-time " . $options[49]['value'] . ";\n
-\tmin-lease-time 7200;\n
-\n
-\n		
-\t#######################################\n
-\t# USERS\n
+				$data .= "\tdefault-lease-time " . $options[49]['value'] . ";
+\tmax-lease-time " . $options[49]['value'] . ";
+\tmin-lease-time 7200;
+	
+\t#######################################
+\t# USERS
 \t#######################################\n\n";
 				
 				$modelsIp = Ip::find()->joinWith('modelDevice')->where(['subnet' => $dhcpSubnet->id, 'type' => 5])->orderBy('ip')->all();
 				foreach ($modelsIp as $modelIp){
-					$data .= "\thost " . $modelIp->modelDevice->id . " {\n
-\t\thardware ethernet " .$modelIp->modelDevice->mac . ";\n
-\t\tfixed-address " . $modelIp->ip . ";\n
+					$data .= "\thost " . $modelIp->modelDevice->id . " {
+\t\thardware ethernet " .$modelIp->modelDevice->mac . ";
+\t\tfixed-address " . $modelIp->ip . ";
 \t}\n";
 					//var_dump($modelIP->modelDevice);
 				}
 				$data .= "}";
 // 				exit();
-				$fileConfig = $test . '/subnets/' . iconv('UTF-8', 'ISO-8859-1//TRANSLIT//IGNORE', $dhcpSubnet->desc) . '.conf';
+				$sysout = system('rm ' . $test . '/subnets/'  . $dhcpSubnet->id . '.conf');
+				
+				$fileConfig = $test . '/subnets/' . $dhcpSubnet->id . '.conf';
 				$file = fopen($fileConfig, 'w');
 				fwrite($file, $data);
 				fclose($file);
